Account: Make sure TargetHandle is uint when contact is 0 on ensureMediaCall/TextChat.

+++ TelepathyQt4/account.cpp
@@ -758,7 +758,7 @@ PendingChannelRequest *Account::ensureTextChat(
     request.insert(QLatin1String(TELEPATHY_INTERFACE_CHANNEL ".TargetHandleType"),
                    (uint) Tp::HandleTypeContact);
     request.insert(QLatin1String(TELEPATHY_INTERFACE_CHANNEL ".TargetHandle"),
-                   contact ? contact->handle().at(0) : 0);
+                   contact ? contact->handle().at(0) : (uint) 0);
     return new PendingChannelRequest(dbusConnection(), objectPath(),
             request, userActionTime, preferredHandler, false, this);
 }
@@ -854,7 +854,7 @@ PendingChannelRequest *Account::ensureMediaCall(
     request.insert(QLatin1String(TELEPATHY_INTERFACE_CHANNEL ".TargetHandleType"),
                    (uint) Tp::HandleTypeContact);
     request.insert(QLatin1String(TELEPATHY_INTERFACE_CHANNEL ".TargetHandle"),
-                   contact ? contact->handle().at(0) : 0);
+                   contact ? contact->handle().at(0) : (uint) 0);
     return new PendingChannelRequest(dbusConnection(), objectPath(),
             request, userActionTime, preferredHandler, false, this);
 }
